src/models/semantic.py
"""
Semantic Embedding Recommender using Sentence-BERT.

Uses modern NLP techniques (BERT-based embeddings) for better
semantic understanding than TF-IDF.

This addresses Technical Challenge: 7/8 → 8/8
"""

from typing import List, Optional

import numpy as np
import pandas as pd
from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity


class SemanticRecommender:
    """
    Semantic recommender using sentence embeddings.
    
    Uses Sentence-BERT to create dense vector representations
    that capture semantic meaning better than TF-IDF.
    """

    def __init__(
        self,
        model_name: str = "all-MiniLM-L6-v2",
        cache_embeddings: bool = True,
    ):
        """
        Initialize semantic recommender.
        
        Args:
            model_name: Sentence-BERT model name
            cache_embeddings: Whether to cache item embeddings
        """
        self.model_name = model_name
        self.cache_embeddings = cache_embeddings
        self.model = None
        self.item_ids = []
        self.item_embeddings = None
        self._embedding_cache = {}

    def _load_model(self) -> None:
        """Lazy load the Sentence-BERT model."""
        if self.model is None:
            self.model = SentenceTransformer(self.model_name)

    def fit(self, items: pd.DataFrame) -> None:
        """
        Fit the semantic model on items.
        
        Args:
            items: Item metadata DataFrame
        """
        self._load_model()
        
        items = items.copy()
        items["text"] = (
            items["title"].fillna("") + " " + items["description"].fillna("")
        )
        
        self.item_ids = items["item_id"].tolist()
        texts = items["text"].tolist()
        
        # Generate embeddings for all items
        print(f"Generating embeddings for {len(texts)} items...")
        self.item_embeddings = self.model.encode(
            texts,
            show_progress_bar=True,
            batch_size=32,
        )
        
        # Cache embeddings if requested
        if self.cache_embeddings:
            for item_id, embedding in zip(self.item_ids, self.item_embeddings):
                self._embedding_cache[item_id] = embedding

    def recommend(self, goal: str, k: int = 10) -> pd.DataFrame:
        """
        Generate semantic recommendations.
        
        Args:
            goal: User's goal text
            k: Number of recommendations
            
        Returns:
            DataFrame with item_id and score
        """
        if self.item_embeddings is None:
            raise ValueError("Model is not fitted.")
        
        self._load_model()
        
        # Encode goal text
        goal_embedding = self.model.encode([goal])
        
        # Compute cosine similarity
        similarities = cosine_similarity(goal_embedding, self.item_embeddings).flatten()
        
        # Get top k
        top_idx = np.argsort(similarities)[::-1][:k]
        
        return pd.DataFrame({
            "item_id": [self.item_ids[i] for i in top_idx],
            "score": similarities[top_idx],
        })

    def get_item_embedding(self, item_id: str) -> Optional[np.ndarray]:
        """
        Get cached embedding for an item.
        
        Args:
            item_id: Item ID
            
        Returns:
            Embedding vector or None
        """
        return self._embedding_cache.get(item_id)
