requirements.txt
pandas
numpy
scikit-learn
scipy
streamlit
matplotlib
pyyaml
pytest
statsmodels
lightgbm
sentence-transformers
shap
plotly

